split smart selectors to atom and group
